[Dharmendra/vikash] Diveded server.js into server.js and routes.js
What changes have you made?

Refactored the codes written in server.js and all the logic part and
framework part are diveded into routes.js and server.js respectively.
implemented events emitter to call the respective handler function for
particular request.

Why do we need this change?

We have to invoke  defferent handlers for different request separately but
After implementation of event we don't have to invoke each handeler separately
for any client request.

What can this change impact?
In further development if their will be a need of implementation of another handler,
we will have to just add that handler name into handlers object and after we don't have
to worry about passing the parameter to the handler

*