[Bindu/Abhishek] - Implemented shoot,hit and miss events
What changes have you made?
First,we have modified package.json for watcher.
We have created shoot function, hit and miss events.
Created Destroy function.
Added tests for all of them.
*

Why do we need this change?
In the game,player should have an ability to shoot the opponent player's
fleet.
The shoot can be a hit or a miss thats why we need a hit and miss event to
inform the player that your respective shot was a hit or a miss.
A hit will cause a player loosing a ship's weak point and a miss will be
just for the information to the opponent player.
The destroy function will increment the hitted holes of the respective ship
after a hit.
*

What can this change impact?
This will make a huge Impact to player.
He can play in turn wise and shoot the opponent player's fleet.
There are hit and miss events to tell the status of his last hit or miss.
*