[Dharmendra/Arun] Added serve_ship_deployment_info function in routes.js
What changes have you made?

Added a function into serve_ship_deployment_info routes.js. And accordingly
modified the  reply_to_deployment function into fetch_updates.js to get the
information. Modified shipTest.js and battleship.js

Why do we need this change?

we have to serve the information regarding any deployment of ship to the player
whether the position are valid or not, and after that player can move further
to deploy his ship or start the game.
 And also modified the reply_to_deployment and reply_to_deployment function into
fetch_updates.js to generate client request and send the information about which
ship player wants to diploy his ship and in which position of opponent player
he wants to shoot to the server. And depending on the server response these function
will update the user interface.

What can this change impact?
In further stages it will handle all the client side request regarding deployment of
ship and a shoot event. after that it will updates the information into user interface.

*