[Vikas/Bindu] - implemented sunk function and checked gameover condition
What changes have you made?
We have written sunk function.
In hit event,gameover condition is implemented.
wrote tests for sunk function.
*

Why do we need this change?
we have to check the ship is sunk or not in each hit to check game over
for the announcing who won.
*

What can this change impact?
sunk function checks ship is sunk or not.If ship is sunk and total
used position is zero, then game is over ;otherwise observer will give
 turn to opponent player.
If ship is not sunk then also observer will give turn to
opponent player.
*