Abhishek -Change the Name Observer to Game
What changes have you made?

Change the Name of 'obsever' to 'game'.
One of the test was failing due to a error make it pased.

Why do we need this change?

We realise that an observer can be also another player which can watch
the game.
So we decided to rename our entity to 'game'
It has all his previous behaviour and rules of the game

What can this change impact?

The working of the game will not effect.
It was simply conceptual thing that we missed earlier.