[Abhishek/Vikas] - Fixed the server,removed a bug,added addPlayer Method.
What changes have you made?
Added a error event in server while handling the request,
Added a method that can respond a user and create a Player on the Server.
Added cookie for tracking the User and giving him a id that can be used
by client side javascript for making the request.

Why do we need this change?
We created a cookie to track the player and give him update according to
his identity.
Server was carshing while the port was already in use now we are catching
that error.

What can this change impact?
This will add a ability to server to response different user simultaneously.
also it will resolve the problem for adding and differentiating between players.