[Dharmendra/Arun] Added two function in fetch_updates and a function in routes.js to validate the shoot.
What changes have you made?

We had added a function sayReady in fetch_updates.js to allow the player
to say ready as soon as he have finished his ship deployment, and added a
function i_am_ready in routes.js to handle the player request i'm ready,
also we have added a function validateShoot in routes.js to handle the player
shoot call, and added a function get_opponentPlayer_id in same file routes.js

Why do we need this change?
Each player has to say i'm ready so that game can decide that who will play
first, so to full fill this need we have created that function sayReady it
will generate a xmlHTTPrequest to talk to the server. Also game has to validate
the shoot shot by the player whether it's a hit or miss.

What can this change impact?

Now player is able to shoot on his opponent territory.