[Dharmendra/Abhishek]-- Implemented updates Feature.
What changes have you made?
We have just made a request from the client to ask for the latest updates
from the server and implemented the method on the server to handle it.

Why do we need this change?
A player wants to know the update of his all the fleet and which are
destroyed.So he can ask to server for the changes on his all remaining
positions.

What can this change impact?
This will give a real view to the player that he is playing the game and
his ships are getting destroyed.