[Abhishek/Vikas] -Changed the client side js using JQuery
What changes have you made?
We have made all the request with JQuery and now the behaviour is as usual
and it is working on another browser also.
We have also added a method for gameEnd

Why do we need this change?
Changing js in JQuery library reduce the line of code and make it easier.
It also resolve the browser dependency.