[Arun/Abhishek]- Notify the player for his turn.
What changes have you made?
We have added new fields in fetch-updates to get the game's status.
Now server is also telling the player who's turn is this.
Server is also telling if you lost the game.

Why do we need this change?
Server needs to Notify the player's turn to the respective player
on the basis of that player can play the game.

What can this change impact?

This will give more information to the player when he should play his moves.