[Bindu/Vikas] - Provided information about hit and sunk.
What changes have you made?
We have provided the information about the ships which got hit and sunk.
It will update each second and give the player the information about hit
ships and which ship is sunk.
*

Why do we need this change?
A player wants to know the status of his ship ,ie.,which ship got hit and
which ship is sunk.
*

What can this change impact?
This will give the player a clear view of his fleet.
*