[Arun/Abhishek] - Splited the UI part into 'deploy' and 'shooting'.
What changes have you made?
We have created new files deploy.html,deploy.css,deploy.js
and modified routes.js for handling the request and redirecting
to deploy.html and then battleship.html in the sequence.
We modified the fetch_updates.js in two parts and made deploy.js

Why do we need this change?
For keeping the logic seperate of deploying the Ship and shooting
we had splitted these parts in to two parts.

What can this change impact?
If the user refresh in the middle of the game the harbour section
won't display in middle.