[vikas/bindu] - Added game over functionality.
What changes have you made?
We have added a property called isAlive in player.According to that,we are
giving the player whether he has won or lost.If the game over,then the
updations will end and all cookies will be deleted.

*

Why do we need this change?
If game is over,player must able to know the information ie.,lost or win.
After that he can't shoot.
*

What can this change impact?
Player will get a clear view of his status in game.
*