[Abhishek] - Re-factored the code of routes.js
What changes have you made?
The methods which were repeating many times like
authentication and dataparsing, now they are done on
every request.I have made seperate file for them
and also have written the test for it.
Because this methods are executed every time and
they should be tested.

Why do we need this change?
To reduce the size of the source and split it into
modules.So it will be easy to add a new handler.

What can this change impact?
This change will do authenication on every request
and add playerId and data to request itself.
That can be easily accesible in every handler.