[Abhishek] - Fix the the logic.
What changes have you made?
I have moved one logical part of the game from
routes.js to our library.(The changes are of storing
hits and miss of other player.)

Why do we need this change?
We realized that this logic can be tested and it sholud be
in logic part not in the routes.js .

What can this change impact?
This changes are not going to change  any massive behavioural
change.It can be treated as refatoring of code.