[Dharmendra/arun] Changed the UI and created an file log.js to log the game information
What changes have you made?

Added a file log.js to log the players and any error information. Changed
the ui part of displaying the game over message and providing the facility
to restart and quit the game. Also modified the function getCookie in
cookiehandlers.js file.

Why do we need this change?

We need some short of information to debug our app when something will
go wrong, so we added a function to log all important information. Changed
the ui part of displaying the game over message and providing the facility
to restart and quit the game because we got some feedback regarding it,
that it's not necessary to display the grid and other information when game
will finished. We just have to provide that you won or loss and provide
the restart and quit facility.

What can this change impact?

*