[Dharmendra] Fixed the bug which display the 	turn message after game over.
What changes have you made?

Fixed that bug which was displaying the turn message after game over. In
order to fixed it, I'd written code to delete that message after game over.

Why do we need this change?

There is nothing worthy to tell the player that this is your turn when game
has over. So this is likely to a bug, which has to be fixed.

What can this change impact?

Now it won't show that turn message after game over.