[Abhishek/Dharmendra] --Injected players object  to routers.
What changes have you made?
We have just moved the global players object to routers
in which we can inject any conditions of the game.
For live game we will provide the object from
server which can be edited.

Why do we need this change?
This dependency can be injected throughly and test
can be written according and they can be independent
because any way we can inject players object in test.

What can this change impact?

*