[Dharmenra/Abhishek] --Moved the validation behaviour to player.
What changes have you made?
We have changed the some game behaviour to the player.
So player can independently place his ships without help
of another entity.

Why do we need this change?
we want to make player entity independent so we have decided
to make this change.

What can this change impact?
This can change the behaviour of game in future, so we have decided
to keep the older functionality, later we can switch on this new
 behaiour.