[Arun/Dharam] - added two methods gameStatus and getUpdates in game.js
What changes have you made?
We have added two method gameStatus and getUpdates in game.js and wrote
some test for these two method and destroy method of game object.

Why do we need this change?
Player wants to know that the what is the game current status, and wants
latest updates of the game. Earlier we were doing all these operation
in routers it self. now we wants to make our router simple so that we have
move this responsibility of serving updates to game.

What can this change impact?
It will be easy to test routers independently. It will reduce one logic from
the routers.