[Dharmendra] Added a method gotHit in ship object
What changes have you made?

Added a method gotHit in ship object to change the vanishedLives whenever
ship got a hit, it will return the ship name if ship got hit. And wrote a
test for this method and modified the game object and its test according
to the new method gothit of ship object

Why do we need this change?

we don't want to let any out sider object to access and modified the ship
data.

What can this change impact?

*