[Vikas upadhyay]
What changes have you made?

for deploying ship previously we are taking all position of ship but now
we are taking initial position of ship and alignment of ship.It will generate
all position of ship according to size of ship and alignment.
for doing this we have created a function name as getAllPositionOfShip.