[Abhishek] - Added Module for bot Player.
What changes have you made?
I have added two lib files and one test file.
One of the library has all http calles for the
bot player.Second file GridPriority has all the
logic for seeting the priority of the grid.

Why do we need this change?
We want the to facilitate the player to play with
computer.

What can this change impact?
This is expermental branch so overall
nothing is really gone change.