[Dharam]  Improved the the way of displaying the messages to player
What changes have you made?

Uses the background colors for message div initially when there is no opponent
player or opponent is not ready. And once the opponent is ready that message
won't be displayed. Instead of displaying the turn message we are just
discreasing the opacity of target grid and keep it back as soon as it's player turn